Makefile,spec: don't compress the man-page
Package managers like to compress man-pages on their own,
because the type of compression for man-pages is user-configurable.
In particular, Gentoo [1] doesn't want packages to compress their man-pages.
Gentoo Portage has workarounds for this,
but this is not specified in Package Manager Specification and
results in extra compression-decompression pass.

RPM also compresses man-pages itself (in `brp-compress`)
(and similarly recompresses them as needed)
rather than relying on packages to install compressed man-pages.

Automake can handle installation of man-pages without the explicit "install" target,
so use the standard automake-provided way of installing man-pages.
It's also smart enough to package `nload.1.in` automatically.

Don't specify an explicit man-page extension in .spec file as recommended by Fedora.

[1] https://github.com/gentoo/gentoo/pull/9543#issuecomment-415662844