Move to 'block' characters for bar graph display
This change modifies the characters used to display the "progressbars"
of the graph. Instead of being purely ascii characters such as `#`, `|`,
and `.`, it uses longer 'block elements' characters `█`, `▟`, and `▗`,
respectively. On terminals which support it, this makes for a much
improved visual display.

This is accomplished by swapping the characters being printed, then by
changing the program/build system slightly so that it will link against
`ncursesw` (the version of ncurses which supports full-width characters
instead of only purely ascii).

Since this change would permanently shift this entire codebase so that
it uses ncursesw instead of "classic ncurses", and since it would
_force_ the use of multiwidth 'block' characters which may not be
compatible with many of the original terminals where nload is used, I
don't recommend upstreaming this change to nload, at least not until
someone makes further changes which make this a toggle-able setting and
which maintains total backwards compatibility with systems where
ncursesw isn't available.

Until then, this change will exist as an option for those who want it
(like me), and as an example for the interested.